+++ package.json
@@ -21,7 +21,9 @@
     "bcrypt": "^5.0.1",
     "body-parser": "^1.20.2",
     "cookie-parser": "^1.4.6",
+    "dotenv": "^16.3.1",
     "ejs": "^3.1.9",
+    "env": "^0.0.2",
     "exceljs": "^4.4.0",
     "express": "^4.18.2",
     "express-mysql-session": "^3.0.0",
